Add version command
This commit adds a version command to `kci` that returns a simple
BuildTime that is injected during the build process.

This is primarily so I can see what I've installed locally, so I'm not
going full into a SemVer build scheme that I'll constantly forget to
update. We can add that, and the automation around it, when it becomes
important.